backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
openai>=1.0
pydantic>=2.0
opencv-python>=4.8
pillow>=10.0
pymupdf>=1.23
python-docx>=1.0
imagehash>=4.3
